refactor(product): split session slot synchronization
