+++ apps/packages/product-client/src/hooks/sessions/workflows/use-session-selection-actions.ts
@@ -7,7 +7,6 @@ import {
   type SessionActivationOutcome,
 } from "#product/hooks/sessions/workflows/session-activation-guard";
 import type { SessionLatencyFlowOptions } from "#product/hooks/sessions/workflows/session-selection-options";
-import { resolveStatusFromExecutionSummary } from "#product/domain/sessions/activity";
 import { isHotReopenEligibleSessionSlot } from "#product/lib/domain/workspaces/selection/hot-reopen";
 import { resolveWorkspaceUiKey } from "#product/lib/domain/workspaces/selection/workspace-ui-key";
 import {
@@ -19,7 +18,6 @@ import {
   finishOrCancelMeasurementOperation,
   finishMeasurementOperation,
   markOperationForNextCommit,
-  recordMeasurementMetric,
   recordMeasurementWorkflowStep,
   startMeasurementOperation,
 } from "#product/lib/infra/measurement/measurement-port";
@@ -28,12 +26,9 @@ import { annotateLatencyFlow } from "#product/lib/infra/measurement/measurement-
 import { scheduleAfterNextPaint } from "#product/lib/infra/scheduling/schedule-after-next-paint";
 import { rememberLastViewedSession } from "#product/stores/preferences/workspace-ui-store";
 import {
-  createEmptySessionRecord,
   findClientSessionIdByMaterializedSessionId,
   getSessionRecord,
   isPendingSessionId,
-  patchSessionRecord,
-  putSessionRecord,
 } from "#product/stores/sessions/session-records";
 import { useSessionSelectionStore } from "#product/stores/sessions/session-selection-store";
 import { useWorkspaceRuntimeBlock } from "#product/hooks/workspaces/derived/use-workspace-runtime-block";
@@ -42,6 +37,7 @@ import {
   SESSION_HOT_SWITCH_MEASUREMENT_SURFACES,
   SESSION_SWITCH_MEASUREMENT_SURFACES,
 } from "#product/hooks/sessions/workflows/session-selection-measurement";
+import { syncSessionSelectionSlot } from "#product/hooks/sessions/workflows/session-selection-slot-sync";
 
 interface UseSessionSelectionWorkflowActionsOptions {
   activateSession: (sessionId: string) => void;
@@ -308,80 +304,15 @@ export function useSessionSelectionWorkflowActions({
     const metadataSessionId = existingSlot?.materializedSessionId
       ?? (sessionId === requestedSessionId ? sessionId : requestedSessionId);
     const sessionMeta = sessions.find((session) => session.id === metadataSessionId) ?? null;
-    const agentKind = existingSlot?.agentKind ?? sessionMeta?.agentKind ?? "unknown";
-
-    if (!existingSlot) {
-      const storeStartedAt = performance.now();
-      putSessionRecord({
-        ...createEmptySessionRecord(sessionId, agentKind, {
-          workspaceId,
-          materializedSessionId: metadataSessionId,
-          modelId: sessionMeta?.modelId ?? null,
-          requestedModelId: sessionMeta?.requestedModelId ?? sessionMeta?.modelId ?? null,
-          modeId: sessionMeta?.modeId ?? null,
-          title: sessionMeta?.title ?? null,
-          liveConfig: sessionMeta?.liveConfig ?? null,
-          executionSummary: sessionMeta?.executionSummary ?? null,
-          mcpBindingSummaries: sessionMeta?.mcpBindingSummaries ?? null,
-          lastPromptAt: sessionMeta?.lastPromptAt ?? null,
-          sessionRelationship: sessionSelectionRelationship,
-        }),
-        status: resolveStatusFromExecutionSummary(
-          sessionMeta?.executionSummary ?? null,
-          sessionMeta?.status ?? "idle",
-        ),
-      });
-      if (measurementOperationId) {
-        recordMeasurementMetric({
-          type: "store",
-          category: "session.list",
-          operationId: measurementOperationId,
-          durationMs: performance.now() - storeStartedAt,
-        });
-      }
-      recordMeasurementWorkflowStep({
-        operationId: measurementOperationId,
-        step: "session.select.slot_store",
-        startedAt: storeStartedAt,
-        outcome: "cache_miss",
-      });
-    } else {
-      const storeStartedAt = performance.now();
-      patchSessionRecord(sessionId, {
-        workspaceId,
-        agentKind,
-        modelId: sessionMeta?.modelId ?? existingSlot.modelId ?? null,
-        requestedModelId:
-          sessionMeta?.requestedModelId
-          ?? sessionMeta?.modelId
-          ?? existingSlot.requestedModelId
-          ?? null,
-        modeId: sessionMeta?.modeId ?? existingSlot.modeId ?? null,
-        title: sessionMeta?.title ?? existingSlot.title ?? null,
-        liveConfig: sessionMeta?.liveConfig ?? existingSlot.liveConfig ?? null,
-        executionSummary: sessionMeta?.executionSummary ?? existingSlot.executionSummary ?? null,
-        mcpBindingSummaries: sessionMeta?.mcpBindingSummaries ?? existingSlot.mcpBindingSummaries ?? null,
-        status: resolveStatusFromExecutionSummary(
-          sessionMeta?.executionSummary ?? existingSlot.executionSummary ?? null,
-          sessionMeta?.status ?? existingSlot.status,
-        ),
-        lastPromptAt: sessionMeta?.lastPromptAt ?? existingSlot.lastPromptAt ?? null,
-      });
-      if (measurementOperationId) {
-        recordMeasurementMetric({
-          type: "store",
-          category: "session.list",
-          operationId: measurementOperationId,
-          durationMs: performance.now() - storeStartedAt,
-        });
-      }
-      recordMeasurementWorkflowStep({
-        operationId: measurementOperationId,
-        step: "session.select.slot_store",
-        startedAt: storeStartedAt,
-        outcome: "cache_hit",
-      });
-    }
+    syncSessionSelectionSlot({
+      existingSlot,
+      materializedSessionId: metadataSessionId,
+      measurementOperationId,
+      sessionId,
+      sessionMeta,
+      sessionRelationship: sessionSelectionRelationship,
+      workspaceId,
+    });
 
     if (committedSessionId !== sessionId) {
       const commitOutcome = commitSelection();
