+++ apps/packages/product-client/src/hooks/workspaces/workflows/use-workspace-bootstrap-actions.ts
@@ -27,15 +27,9 @@ import {
 } from "#product/lib/infra/measurement/measurement-port";
 import type { MeasurementFinishReason } from "#product/lib/domain/telemetry/debug-measurement-catalog";
 import { useUserPreferencesStore } from "#product/stores/preferences/user-preferences-store";
+import { clearLastViewedSession, useWorkspaceUiStore } from "#product/stores/preferences/workspace-ui-store";
 import {
-  clearLastViewedSession,
-  useWorkspaceUiStore,
-} from "#product/stores/preferences/workspace-ui-store";
-import {
-  findClientSessionIdByMaterializedSessionId,
-  getSessionRecord,
-  patchSessionRecord,
-  removeSessionRecord,
+  findClientSessionIdByMaterializedSessionId, getSessionRecord, patchSessionRecord, removeSessionRecord,
 } from "#product/stores/sessions/session-records";
 import { useSessionSelectionStore } from "#product/stores/sessions/session-selection-store";
 import { markWorkspaceBootstrappedInSession } from "#product/hooks/workspaces/lifecycle/workspace-bootstrap-memory";
